♻️ move to scikit-build-core and src-layout
