--- MANIFEST.in
@@ -1,48 +1,0 @@
-include README.md LICENSE CMakeLists.txt
-include src/*
-include include/*
-include apps/*
-graft mqt/*
-global-exclude __pycache__/
-global-exclude *.py[cod]
-global-exclude *.so
-
-# Include relevant files from other MQT projects
-include extern/mqt-core/CMakeLists.txt
-graft   extern/mqt-core/include
-graft   extern/mqt-core/src
-graft   extern/mqt-core/cmake
-
-# Include necessary files submodules
-graft   extern/mqt-core/extern/json
-prune   extern/mqt-core/extern/json/benchmarks
-prune   extern/mqt-core/extern/json/tools
-prune   extern/mqt-core/extern/json/third_party
-prune   extern/mqt-core/extern/json/include
-
-graft   extern/mqt-core/extern/pybind11
-graft   extern/mqt-core/extern/pybind11_json
-
-graft   extern/mqt-core/extern/boost/config
-graft   extern/mqt-core/extern/boost/multiprecision
-
-graft extern/taskflow
-prune extern/taskflow/image
-prune extern/taskflow/benchmarks
-prune extern/taskflow/doxygen
-prune extern/taskflow/3rd-party
-prune extern/taskflow/tfprof
-prune extern/taskflow/sandbox
-prune extern/taskflow/unittests
-prune extern/taskflow/examples
-
-graft extern/cxxopts
-
-
-prune **/.github
-prune **/doc
-prune **/docs
-prune **/meta
-prune **/plots
-prune **/test
-prune **/tests
