+++ src/CMakeLists.txt
@@ -1,3 +1,7 @@
+# add MQT::Core target
+set(BUILD_MQT_CORE_TESTS
+    OFF
+    CACHE BOOL "Build MQT Core tests")
 add_subdirectory("${PROJECT_SOURCE_DIR}/extern/mqt-core" "extern/mqt-core"
                  EXCLUDE_FROM_ALL)
 
@@ -23,14 +27,12 @@ add_library(
   ${CMAKE_CURRENT_SOURCE_DIR}/UnitarySimulator.cpp
   ${PROJECT_SOURCE_DIR}/include/PathSimulator.hpp
   ${CMAKE_CURRENT_SOURCE_DIR}/PathSimulator.cpp)
-target_include_directories(
-  ${PROJECT_NAME}
-  PUBLIC $<BUILD_INTERFACE:${${PROJECT_NAME}_SOURCE_DIR}/include>)
 
-# cmake-lint: disable=C0103
-set(JSON_MultipleHeaders
-    OFF
-    CACHE INTERNAL "")
+# set include directories
+target_include_directories(${PROJECT_NAME} PUBLIC ${PROJECT_SOURCE_DIR}/include
+                                                  ${PROJECT_BINARY_DIR}/include)
+
+# link to the MQT::Core libraries
 target_link_libraries(${PROJECT_NAME} PUBLIC MQT::CoreDD)
 
 set(TF_BUILD_TESTS
@@ -50,4 +52,9 @@ set_target_properties(
   Taskflow PROPERTIES INTERFACE_SYSTEM_INCLUDE_DIRECTORIES
                       $<TARGET_PROPERTY:Taskflow,INTERFACE_INCLUDE_DIRECTORIES>)
 
+# add MQT alias
 add_library(MQT::${PROJECT_NAME} ALIAS ${PROJECT_NAME})
+
+if(BUILD_MQT_DDSIM_BINDINGS)
+  add_subdirectory(python)
+endif()
