+++ src/mqt/ddsim/_version.pyi
@@ -1,0 +1,4 @@
+__version__: str
+version: str
+__version_tuple__: tuple[int, int, int, str, str] | tuple[int, int, int]
+version_tuple: tuple[int, int, int, str, str] | tuple[int, int, int]
