+++ src/mqt/ddsim/error.py
